HSEARCH-3324 Move ITs related to DocumentElement to a single package
